requirements.txt
tensorflow==2.18.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pillow==10.1.0
numpy==2.0.0 